business logic to controller

+++ src/main/java/com/example/cms/service/ContactService.java
@@ -1,15 +1,18 @@
 package com.example.cms.service;
 
+import com.example.cms.constants.MessageCode;
 import com.example.cms.dto.ContactDTO;
 import com.example.cms.dto.PageDTO;
 import com.example.cms.entity.Contact;
 import com.example.cms.repository.ContactRepository;
 import com.example.cms.utils.ModalMapperUtil;
+import com.example.cms.utils.ResponseHandler;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,15 +24,22 @@ public class ContactService {
 
     @Autowired
     private ContactRepository contactRepository;
+    @Autowired
+    private ResponseHandler responseHandler;
 
-    public ContactDTO saveContact(ContactDTO contactDTO) {
+    public ResponseEntity<Object> saveContact(ContactDTO contactDTO) {
+        contactDTO.setEmail(contactDTO.getEmail().trim().toLowerCase());
+        ContactDTO contactResponse;
+        if (existContactByEmail(contactDTO.getEmail())) {
+            return responseHandler.generateResponse("", MessageCode.CONTACT_EXIST_EMAIL, false, HttpStatus.BAD_REQUEST);
+        }
         Contact contact = new Contact();
         contact.setFirstName(contactDTO.getFirstName());
         contact.setLastName(contactDTO.getLastName());
         contact.setEmail(contactDTO.getEmail());
         contact.setPhoneNumber(contactDTO.getPhoneNumber());
         contact = contactRepository.save(contact);
-        return ModalMapperUtil.map(contact, ContactDTO.class);
+        return responseHandler.generateResponse(ModalMapperUtil.map(contact, ContactDTO.class), MessageCode.CONTACT_CREATED, true, HttpStatus.OK);
     }
 
     public Contact findContact(long id) {
@@ -37,25 +47,44 @@ public class ContactService {
         return optionalContact.orElse(null);
     }
 
-    public ContactDTO editContact(Contact contact, ContactDTO contactDTO) {
-        contact.setFirstName(contactDTO.getFirstName());
-        contact.setLastName(contactDTO.getLastName());
-        contact.setEmail(contactDTO.getEmail());
-        contact.setPhoneNumber(contactDTO.getPhoneNumber());
-        contact = contactRepository.save(contact);
-        return ModalMapperUtil.map(contact, ContactDTO.class);
+    public ResponseEntity<Object> editContact(ContactDTO contactDTO) {
+        contactDTO.setEmail(contactDTO.getEmail().trim().toLowerCase());
+        Contact contact = findContact(contactDTO.getId());
+        if (contact!=null) {
+            if (existContactByEmailAndNotId(contactDTO.getEmail(), contact.getId())) {
+                return responseHandler.generateResponse("", MessageCode.CONTACT_EXIST_EMAIL, false, HttpStatus.BAD_REQUEST);
+            }
+            contact.setFirstName(contactDTO.getFirstName());
+            contact.setLastName(contactDTO.getLastName());
+            contact.setEmail(contactDTO.getEmail());
+            contact.setPhoneNumber(contactDTO.getPhoneNumber());
+            contact = contactRepository.save(contact);
+            return responseHandler.generateResponse(ModalMapperUtil.map(contact, ContactDTO.class), MessageCode.CONTACT_EDITED, true, HttpStatus.OK);
+        }
+        return responseHandler.generateResponse("", MessageCode.CONTACT_NOT_FOUND, false, HttpStatus.BAD_REQUEST);
     }
 
-    public void deleteContact(Contact contact) {
-        contactRepository.delete(contact);
+    public ResponseEntity<Object> deleteContact(long id) {
+        Contact contact = findContact(id);
+        if (contact!=null) {
+            contactRepository.delete(contact);
+            return responseHandler.generateResponse("", MessageCode.CONTACT_DELETED, true, HttpStatus.OK);
+        }
+        return responseHandler.generateResponse("", MessageCode.CONTACT_NOT_FOUND, false, HttpStatus.BAD_REQUEST);
+
     }
 
-    public ContactDTO findContactData(long id) {
-        return ModalMapperUtil.map(findContact(id), ContactDTO.class);
+    public ResponseEntity<Object> findContactData(long id) {
+        Contact contact = findContact(id);
+        if (contact!=null) {
+            return responseHandler.generateResponse(ModalMapperUtil.map(contact, ContactDTO.class), MessageCode.CONTACT_FETCHED, true, HttpStatus.OK);
+        }
+        return responseHandler.generateResponse("", MessageCode.CONTACT_NOT_FOUND, false, HttpStatus.BAD_REQUEST);
     }
 
-    public List<ContactDTO> searchContacts(String keyword) {
-        return ModalMapperUtil.mapAll(contactRepository.findByFirstNameOrLastNameOrEmail(keyword), ContactDTO.class);
+    public ResponseEntity<Object> searchContacts(String keyword) {
+        List<ContactDTO> contactDTOList = ModalMapperUtil.mapAll(contactRepository.findByFirstNameOrLastNameOrEmail(keyword), ContactDTO.class);
+        return responseHandler.generateResponse(contactDTOList, MessageCode.CONTACT_FETCHED, true, HttpStatus.OK);
     }
 
     public boolean existContactByEmail(String email) {
@@ -66,10 +95,10 @@ public class ContactService {
         return contactRepository.existsByEmailAndNotId(email, id);
     }
 
-    public PageDTO findAllContactData(Integer limit, Integer pageNo) {
-        Pageable pageable = PageRequest.of(pageNo, limit);
-        Page<Contact> contacts = contactRepository.findAll(pageable);
-        return new PageDTO(ModalMapperUtil.mapAll(contacts.getContent(), ContactDTO.class), contacts.getContent().size(),
-                contacts.getTotalPages(), contacts.getTotalElements());
+    public ResponseEntity<Object> findAllContactData(Integer limit, Integer pageNo) {
+        Page<Contact> contacts = contactRepository.findAll(PageRequest.of(pageNo, limit));
+        PageDTO response = new PageDTO(ModalMapperUtil.mapAll(contacts.getContent(), ContactDTO.class),
+                contacts.getContent().size(), contacts.getTotalPages(), contacts.getTotalElements());
+        return responseHandler.generateResponse(response, MessageCode.CONTACT_FETCHED, true, HttpStatus.OK);
     }
 }
